Plastic Testing Laboratory in Chennai
Equipped with the advanced appliances for Plastics Testing Laboratory in Chennai, Maeon Laboratories has focussed on a wide range of testing which are accredited with  National Accreditation Board for Testing and Calibration Laboratories (NABL). 

Maeon Laboratories additionally gives the total extent of examination of polymers, plastics and composites and gives best organizations to the Plastic and bound together endeavors by accepting tasks of plastics materials/things/composites testing as indicated by presumed National and Global models.